+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## To be released, v0.9.2
 
+* Allow to read test cases from a directory containing separated input/output files.
 * bugfix: Fix to print standard error while execution. (reported by @nganhkhoa at [#5](https://github.com/fukamachi/getac/issues/5))
 
 ## 2019-12-31, v0.9.1